backend/requirements.txt
fastapi
uvicorn
python-dotenv
pdfplumber
requests
supabase
resend
python-multipart
pydantic
numpy
tenacity
httpx
openai
redis
sentence-transformers
--extra-index-url https://download.pytorch.org/whl/cpu
torch
pydantic-settings